// moonbowcl/src/app/model/dto/attr/attrCosmoConf.ts
/*
* SPDX-FileCopyrightText: (C) Copyright 2023 C.S.I. Piemonte
*
* SPDX-License-Identifier: EUPL-1.2 */

export class AttrCosmoConf {
    include_allegati: boolean;
    allegato_figlio: boolean;
    idPratica: string;
    codiceTipologia: string;
    oggetto: string;
    codiceIpaEnte: string;
    riassunto: string;
    utenteCreazione: string;
    codiceTipoDocIstanza: string;
    codiceTipoDocAllegato: string;
    codiceTipoDocIntegrazione: string;
    codiceTipoDocIntegrazioneAllegato: string;
}